+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Include pyenv and Python 3.9 to 3.12 in the kitchen sink image
+  ([#232](https://github.com/pulumi/pulumi-docker-containers/pull/232))
+
 - Add per-language versions of the `pulumi/pulumi-python` image
   ([#226](https://github.com/pulumi/pulumi-docker-containers/pull/226))