+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - Add Poetry to Python images ([#240](https://github.com/pulumi/pulumi-docker-containers/pull/240))
 
+## 3.128.0
+
 - Update to debian 12 (bookworm) slim as base image
   ([#236](https://github.com/pulumi/pulumi-docker-containers/pull/236))